Example for a map
Displays the number of residents in 1x1km grids for Berlin in the year 2000. Light purple cells indicate a low number of residents, and dark purple cells indicate a high number. We fixed the color scale for each feature so that it approximately ranges from the first to the ninth decile in all cities with more than 100,000 inhabitants.
The data base of the maps is social security data from the IAB. For data protection reasons, we removed cells with fewer than 20 residents. Blue
areas in the background represent water; green areas, forests; light yellow areas, settlements; solid gray lines, roads; and dashed gray lines, railroads.